+++ resdk/resources/process.py
@@ -68,13 +68,8 @@ class Process(BaseResource):
         #: the heart of process - here the algorithm is defined.
         self.run = None
 
-        BaseResource.__init__(self, slug, id, model_data, resolwe)
+        super(Process, self).__init__(slug, id, model_data, resolwe)
 
     def print_inputs(self):
         """Pretty print input_schema."""
         _print_input_line(self.input_schema, 0)  # pylint: disable=no-member
-
-
-def get_process_id(process):
-    """Return id attribute of the object if it is process, othervise return given value."""
-    return process.id if isinstance(process, Process) else process