+++ resdk/resources/sample.py
@@ -23,7 +23,12 @@ class Sample(BaseCollection):
 
     endpoint = 'sample'
 
-    WRITABLE_FIELDS = ('collections', 'presample') + BaseCollection.WRITABLE_FIELDS
+    WRITABLE_FIELDS = ('presample', ) + BaseCollection.WRITABLE_FIELDS
+
+    #: ids of collections objects (if not hydrated), collections objects (if hydrated)
+    _collections = None
+    #: (lazy loaded) list of data object that belong to sample
+    _data = None
 
     def __init__(self, slug=None, id=None,  # pylint: disable=redefined-builtin
                  model_data=None, resolwe=None, presample=False):
@@ -32,7 +37,30 @@ class Sample(BaseCollection):
 
         self.presample = presample
 
-        BaseCollection.__init__(self, slug, id, model_data, resolwe)
+        super(Sample, self).__init__(slug, id, model_data, resolwe)
+
+    def update(self):
+        """Clear cache and update resource fields from the server."""
+        self._collections = None
+        self._data = None
+
+        super(Sample, self).update()
+
+    @property
+    def collections(self):
+        """Lazy load `collection` objects to which sample belongs."""
+        if not self._collections:
+            self._collections = self.resolwe.collection.filter(sample=self.id)
+
+        return self._collections
+
+    @property
+    def data(self):
+        """Lazy load ``data`` objects belonging to the collection."""
+        if not self._data:
+            self._data = self.resolwe.data.filter(sample=self.id)
+
+        return self._data
 
     def print_annotation(self):
         """Provide annotation data."""
@@ -49,8 +77,3 @@ class Sample(BaseCollection):
             self.logger.info('Moved Sample %s from presmaples to samples', self.id)
         else:
             raise NotImplementedError("Method supports objects in presample endpoint only.")
-
-
-def get_sample_id(sample):
-    """Return id attribute of the object if it is sample, othervise return given value."""
-    return sample.id if isinstance(sample, Sample) else sample