+++ resdk/tests/unit/test_base.py
@@ -10,8 +10,7 @@ import slumber
 
 from mock import patch, MagicMock, call
 
-from resdk.resources.base import BaseResource, get_resource_id
-from resdk.resources import Data
+from resdk.resources.base import BaseResource
 
 
 class TestBaseResource(unittest.TestCase):
@@ -61,7 +60,6 @@ class TestBaseResource(unittest.TestCase):
         base_resource.writable_scalar = '42'
         self.assertEqual(base_resource.writable_scalar, '42')
 
-#    @patch('resdk.resources.base.getattr')
     @patch('resdk.resources.base.BaseResource', spec=True)
     def test_id_http_error(self, base_mock):
         base_mock.configure_mock(endpoint='resource_endpoint')
@@ -235,13 +233,6 @@ class TestBaseMethods(unittest.TestCase):
         base_mock.configure_mock(id=1, slug='a', name='b')
         out = BaseResource.__repr__(base_mock)
         self.assertEqual(out, 'BaseResource <id: 1, slug: \'a\', name: \'b\'>')
-
-    def test_get_resource_id(self):
-        data = Data(id=1, resolwe=MagicMock())
-        data.id = 1  # this is overriden when initialized
-        self.assertEqual(get_resource_id(data), 1)
-
-        self.assertEqual(get_resource_id(2), 2)
 
 
 if __name__ == '__main__':